🏘🐯 Checkpoint ./commands/mods/weebhook.js:23585414/1738

+++ commands/mods/weebhook.js
@@ -1,11 +1,15 @@
 const discord = require("discord.js");
 const embed = new discord.MessageEmbed();
 const db = require("quick.db");
+const { prefix, bowner } = require("../../config.json");
 
 module.exports = {
   name: "mhook",
   description: "",
   async run(client, message, args) {
+    if (message.author.id !== bowner) return;
+    const channel = message.channel;
+
     message.delete();
     message.channel
       .createWebhook("Marvel", {
@@ -13,7 +17,7 @@ module.exports = {
           "https://cdn.discordapp.com/avatars/748583869527097376/8b9300203ff8f2b17e509634c44dfba7.jpeg"
       })
       .then(webhook =>
-        message.channel.send(
+        message.reply(
           "Weebhook Created With Name Marvel By : <@!" + message.author + ">"
         )
       );
